--- a/saleor/graphql/order/types.py
+++ b/saleor/graphql/order/types.py
@@ -438,13 +438,23 @@ class OrderEvent(ModelObjectType[models.OrderEvent]):
     @staticmethod
     def resolve_app(root: models.OrderEvent, info):
         requestor = get_user_or_app_from_context(info.context)
-        check_is_owner_or_has_one_of_perms(
-            requestor,
-            root.user,
-            AppPermission.MANAGE_APPS,
-            OrderPermissions.MANAGE_ORDERS,
-        )
-        return AppByIdLoader(info.context).load(root.app_id) if root.app_id else None
+
+        def _resolve_app(user):
+            check_is_owner_or_has_one_of_perms(
+                requestor,
+                user,
+                AppPermission.MANAGE_APPS,
+                OrderPermissions.MANAGE_ORDERS,
+            )
+            return (
+                AppByIdLoader(info.context).load(root.app_id) if root.app_id else None
+            )
+
+        if root.user_id:
+            return (
+                UserByUserIdLoader(info.context).load(root.user_id).then(_resolve_app)
+            )
+        return _resolve_app(None)
 
     @staticmethod
     def resolve_email(root: models.OrderEvent, _info):