--- a/saleor/graphql/product/dataloaders/products.py
+++ b/saleor/graphql/product/dataloaders/products.py
@@ -471,8 +471,10 @@ class MediaByProductVariantIdLoader(DataLoader):
         )
 
         variant_media_pairs = defaultdict(list)
+        media_ids = set()
         for variant_id, media_id in variant_media.iterator():
             variant_media_pairs[variant_id].append(media_id)
+            media_ids.add(media_id)
 
         def map_variant_media(variant_media):
             media_map = {media.id: media for media in variant_media}
@@ -483,7 +485,7 @@ class MediaByProductVariantIdLoader(DataLoader):
 
         return (
             ProductMediaByIdLoader(self.context)
-            .load_many(set(media_id for variant_id, media_id in variant_media))
+            .load_many(media_ids)
             .then(map_variant_media)
         )